[ISSUE #934]🔥Support producer send rpc message-3🚀
