+++ rocketmq-broker/src/hook/check_before_put_message.rs
@@ -16,7 +16,7 @@
  */
 use std::sync::Arc;
 
-use rocketmq_common::common::message::message_single::MessageExt;
+use rocketmq_common::common::message::message_ext::MessageExt;
 use rocketmq_store::base::message_result::PutMessageResult;
 use rocketmq_store::config::message_store_config::MessageStoreConfig;
 use rocketmq_store::hook::put_message_hook::PutMessageHook;
