+++ rocketmq-broker/src/util/hook_utils.rs
@@ -21,7 +21,7 @@ use std::sync::atomic::Ordering;
 use std::sync::Arc;
 
 use rocketmq_common::common::config::TopicConfig;
-use rocketmq_common::common::message::message_single::MessageExt;
+use rocketmq_common::common::message::message_ext::MessageExt;
 use rocketmq_common::common::message::message_single::MessageExtBrokerInner;
 use rocketmq_common::common::message::MessageConst;
 use rocketmq_common::common::mix_all::RETRY_GROUP_TOPIC_PREFIX;
