+++ rocketmq-common/src/common/message/message_single.rs
@@ -26,6 +26,7 @@ use bytes::BufMut;
 use bytes::Bytes;
 
 use crate::common::hasher::string_hasher::JavaStringHasher;
+use crate::common::message::message_ext::MessageExt;
 use crate::common::message::MessageConst;
 use crate::common::message::MessageTrait;
 use crate::common::message::MessageVersion;
@@ -253,230 +254,6 @@ impl MessageTrait for Message {
     fn as_any_mut(&mut self) -> &mut dyn Any {
         self
     }
-}
-
-#[derive(Clone, Debug)]
-pub struct MessageExt {
-    pub message: Message,
-    pub broker_name: String,
-    pub queue_id: i32,
-    pub store_size: i32,
-    pub queue_offset: i64,
-    pub sys_flag: i32,
-    pub born_timestamp: i64,
-    pub born_host: SocketAddr,
-    pub store_timestamp: i64,
-    pub store_host: SocketAddr,
-    pub msg_id: String,
-    pub commit_log_offset: i64,
-    pub body_crc: u32,
-    pub reconsume_times: i32,
-    pub prepared_transaction_offset: i64,
-}
-
-impl MessageExt {
-    pub fn socket_address_2_byte_buffer(ip: &SocketAddr) -> bytes::Bytes {
-        match ip {
-            SocketAddr::V4(value) => {
-                let mut byte_buffer = bytes::BytesMut::with_capacity(4 + 4);
-                byte_buffer.put_slice(&value.ip().octets());
-                byte_buffer.put_i32(value.port() as i32);
-                byte_buffer.copy_to_bytes(byte_buffer.len())
-            }
-            SocketAddr::V6(value) => {
-                let mut byte_buffer = bytes::BytesMut::with_capacity(16 + 4);
-                byte_buffer.put_slice(&value.ip().octets());
-                byte_buffer.put_i32(value.port() as i32);
-                byte_buffer.copy_to_bytes(byte_buffer.len())
-            }
-        }
-    }
-
-    pub fn born_host_bytes(&self) -> bytes::Bytes {
-        Self::socket_address_2_byte_buffer(&self.born_host)
-    }
-
-    pub fn born_store_bytes(&self) -> bytes::Bytes {
-        Self::socket_address_2_byte_buffer(&self.store_host)
-    }
-
-    pub fn topic(&self) -> &str {
-        self.message.topic()
-    }
-
-    pub fn born_host(&self) -> SocketAddr {
-        self.born_host
-    }
-
-    pub fn store_host(&self) -> SocketAddr {
-        self.store_host
-    }
-
-    pub fn with_born_host_v6_flag(&mut self) {
-        self.sys_flag |= MessageSysFlag::BORNHOST_V6_FLAG;
-    }
-
-    pub fn with_store_host_v6_flag(&mut self) {
-        self.sys_flag |= MessageSysFlag::STOREHOSTADDRESS_V6_FLAG;
-    }
-
-    pub fn body(&self) -> Option<bytes::Bytes> {
-        self.message.body()
-    }
-
-    #[inline]
-    pub fn sys_flag(&self) -> i32 {
-        self.sys_flag
-    }
-
-    #[inline]
-    pub fn body_crc(&self) -> u32 {
-        self.body_crc
-    }
-
-    #[inline]
-    pub fn queue_id(&self) -> i32 {
-        self.queue_id
-    }
-
-    pub fn flag(&self) -> i32 {
-        self.message.flag()
-    }
-
-    pub fn message_inner(&self) -> &Message {
-        &self.message
-    }
-
-    pub fn broker_name(&self) -> &str {
-        &self.broker_name
-    }
-
-    pub fn store_size(&self) -> i32 {
-        self.store_size
-    }
-
-    pub fn queue_offset(&self) -> i64 {
-        self.queue_offset
-    }
-
-    pub fn born_timestamp(&self) -> i64 {
-        self.born_timestamp
-    }
-
-    pub fn store_timestamp(&self) -> i64 {
-        self.store_timestamp
-    }
-
-    pub fn msg_id(&self) -> &str {
-        &self.msg_id
-    }
-
-    pub fn commit_log_offset(&self) -> i64 {
-        self.commit_log_offset
-    }
-
-    pub fn reconsume_times(&self) -> i32 {
-        self.reconsume_times
-    }
-
-    pub fn prepared_transaction_offset(&self) -> i64 {
-        self.prepared_transaction_offset
-    }
-
-    pub fn set_message_inner(&mut self, message_inner: Message) {
-        self.message = message_inner;
-    }
-
-    pub fn set_broker_name(&mut self, broker_name: String) {
-        self.broker_name = broker_name;
-    }
-
-    pub fn set_queue_id(&mut self, queue_id: i32) {
-        self.queue_id = queue_id;
-    }
-
-    pub fn set_store_size(&mut self, store_size: i32) {
-        self.store_size = store_size;
-    }
-
-    pub fn set_queue_offset(&mut self, queue_offset: i64) {
-        self.queue_offset = queue_offset;
-    }
-
-    pub fn set_sys_flag(&mut self, sys_flag: i32) {
-        self.sys_flag = sys_flag;
-    }
-
-    pub fn set_born_timestamp(&mut self, born_timestamp: i64) {
-        self.born_timestamp = born_timestamp;
-    }
-
-    pub fn set_born_host(&mut self, born_host: SocketAddr) {
-        self.born_host = born_host;
-    }
-
-    pub fn set_store_timestamp(&mut self, store_timestamp: i64) {
-        self.store_timestamp = store_timestamp;
-    }
-
-    pub fn set_store_host(&mut self, store_host: SocketAddr) {
-        self.store_host = store_host;
-    }
-
-    pub fn set_msg_id(&mut self, msg_id: String) {
-        self.msg_id = msg_id;
-    }
-
-    pub fn set_commit_log_offset(&mut self, commit_log_offset: i64) {
-        self.commit_log_offset = commit_log_offset;
-    }
-
-    pub fn set_body_crc(&mut self, body_crc: u32) {
-        self.body_crc = body_crc;
-    }
-
-    pub fn set_reconsume_times(&mut self, reconsume_times: i32) {
-        self.reconsume_times = reconsume_times;
-    }
-
-    pub fn set_prepared_transaction_offset(&mut self, prepared_transaction_offset: i64) {
-        self.prepared_transaction_offset = prepared_transaction_offset;
-    }
-
-    pub fn properties(&self) -> &HashMap<String, String> {
-        self.message.properties()
-    }
-
-    pub fn get_tags(&self) -> Option<String> {
-        self.message.get_tags()
-    }
-}
-
-impl Default for MessageExt {
-    fn default() -> Self {
-        Self {
-            message: Default::default(),
-            broker_name: "".to_string(),
-            queue_id: 0,
-            store_size: 0,
-            queue_offset: 0,
-            sys_flag: 0,
-            born_timestamp: 0,
-            born_host: "127.0.0.1:10911".parse().unwrap(),
-            store_timestamp: 0,
-            store_host: "127.0.0.1:10911".parse().unwrap(),
-            msg_id: "".to_string(),
-            commit_log_offset: 0,
-            body_crc: 0,
-            reconsume_times: 0,
-            prepared_transaction_offset: 0,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct MessageClientExt {
-    pub message_ext_inner: MessageExt,
 }
 
 #[derive(Debug, Default)]
